requirements-test.txt
aiodns==3.2.0
aiohttp==3.9.5
aioodbc==0.5.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
azure-ai-documentintelligence==1.0.0b3
azure-ai-language-questionanswering==1.1.0
azure-core==1.31.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-functions==1.20.0
azure-monitor-events-extension==0.1.0
azure-monitor-opentelemetry==1.6.0
azure-monitor-opentelemetry-exporter==1.0.0b27
azure-storage-blob==12.20.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.2.0
cryptography==43.0.0
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
fixedint==0.1.6
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.2
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
isodate==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.23
langchain-mistralai==0.1.10
langchain-openai==0.1.17
langsmith==0.1.93
mammoth==1.8.0
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
openai==1.37.0
opentelemetry-api==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-dbapi==0.46b0
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-psycopg2==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-urllib==0.46b0
opentelemetry-instrumentation-urllib3==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
packaging==24.1
pluggy==1.5.0
psutil==5.9.8
pycares==4.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pyodbc==5.1.0
pytest==7.2.1
pytest-asyncio==0.23.7
pytest-cov==4.0.0
pytest-dotenv==0.5.2
pytest-mock==3.14.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
azure-identity==1.19.0
redis==5.2.1
aiocache==0.12.3
azure-keyvault-secrets==4.9.0